flask_vite_ts/backend/requirements.txt
flask~=3.0.0
flask-cors~=4.0.0
flask-sqlalchemy~=3.1